# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MSPR HealthAI Coach - AI Nutrition",
    "description": "\nMicro-service d'analyse nutritionnelle et de generation de plans repas par IA, partie de la plateforme MSPR HealthAI Coach.\n\nDeux pipelines d'IA sont exposes :\n\n- **Analyse de repas** : photo en entree, classification HuggingFace (`nateraw/food`, modele Food-101), lookup nutritionnel sur les datasets ETL, detection de desequilibres macros vs profil utilisateur, puis recommandations generees par Ollama (Gemma3:4b) avec fallback matrice statique.\n- **Generation de plans repas** : objectif de sante, preferences alimentaires et budget, prompt structure soumis a Ollama, validation du JSON et persistance du plan sur 1 a 30 jours.\n\nLe service expose egalement la gestion du profil nutritionnel (`nutrition-goals/me`) et l'historique pagine des analyses et des plans (`meal-analyses/me`, `meal-plans/me`). L'authentification se fait via JWT signe par MSPR-AUTH (`Authorization: Bearer <jwt>`).\n",
    "contact": {
      "name": "Arthur Poncin",
      "email": "arthur.poncin@sixense-group.com"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/generate-meal-plan": {
      "post": {
        "tags": [
          "Plans"
        ],
        "summary": "Genere un plan repas personnalise (1 a 30 jours)",
        "description": "Genere un plan repas personnalise sur 1 a 30 jours en fonction de l'objectif sante, du regime, des allergies et du budget.\n\nPipeline applique :\n\n1. Resolution de l'objectif sante : valeur explicite > profil utilisateur (`nutrition-goals/me`) > `balance` par defaut.\n2. Lecture des entitlements (tier utilisateur via MSPR-AUTH). Les tiers `premium` et `premium_plus` bypassent le cache et regenerent systematiquement.\n3. Cache contenu : hash canonique des inputs (objectif, regime, allergies triees, budget, calories cible, duree). Cache hit, on renvoie le plan stocke.\n4. Cache miss : appel Ollama (Gemma3:4b) avec prompt structure et `format: json`. En cas d'echec ou JSON invalide, repli sur la matrice 16 plans statiques (`fallback: true`).\n5. Persistance dans `meal_plans` et renvoi du plan complet.\n\n**Rate limit** : 10 requetes par heure et 3 par minute par utilisateur (header standard `Retry-After` en cas de depassement).\n\n**Latence** : 5-30 s sur CPU pour une generation LLM. Cache hit < 100 ms.\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis. Le JWT est aussi propage a MSPR-AUTH pour resoudre les entitlements.",
        "operationId": "generate_meal_plan_api_v1_generate_meal_plan_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealPlanRequest"
              },
              "examples": {
                "muscle_gain_omnivore": {
                  "summary": "Prise de masse, omnivore, 7 jours",
                  "value": {
                    "health_goal": "muscle_gain",
                    "diet_type": "omnivore",
                    "duration_days": 7,
                    "allergies": [],
                    "budget_eur_per_day": 15.0
                  }
                },
                "weight_loss_vegetarien": {
                  "summary": "Perte de poids, vegetarien, 14 jours",
                  "value": {
                    "health_goal": "weight_loss",
                    "diet_type": "vegetarien",
                    "duration_days": 14,
                    "allergies": [
                      "arachides"
                    ],
                    "budget_eur_per_day": 10.0
                  }
                },
                "balance_vegan_minimum": {
                  "summary": "Equilibre, vegan, requete minimale (1 jour, sans budget)",
                  "value": {
                    "diet_type": "vegan",
                    "duration_days": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan repas genere ou recupere du cache. `fallback: true` indique un repli matrice statique.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealPlanResponse"
                },
                "example": {
                  "plan_id": 42,
                  "fallback": false,
                  "days": [
                    {
                      "day": 1,
                      "meals": [
                        {
                          "name": "Bowl quinoa, poulet et legumes verts",
                          "macros": {
                            "calories": 620,
                            "protein_g": 45.0,
                            "carbs_g": 65.0,
                            "fat_g": 18.0
                          },
                          "ingredients": [
                            "100 g quinoa",
                            "150 g blanc de poulet",
                            "200 g brocolis",
                            "1 c.s. huile d'olive"
                          ],
                          "est_budget_eur": 5.5,
                          "prep_time_min": 25
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Payload invalide (regime inconnu, duree hors [1, 30], etc)."
          },
          "503": {
            "description": "Contraintes infaisables (allergies / regime impossibles a satisfaire) ou Ollama injoignable et fallback statique indisponible.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Contraintes infaisables, ajustez vos contraintes (allergies / regime).",
                  "infeasible": true
                }
              }
            }
          },
          "429": {
            "description": "Rate limit depasse (10/heure ou 3/minute)."
          }
        }
      }
    },
    "/api/v1/meal-plans/me": {
      "get": {
        "tags": [
          "Historique"
        ],
        "summary": "Historique pagine des plans repas de l'utilisateur",
        "description": "Renvoie l'historique pagine des plans repas generes pour l'utilisateur authentifie.\n\nTri decroissant par date de generation. Chaque element contient l'objectif sante, les contraintes (regime, allergies, budget, duree, calories cible) et le plan complet (jours et repas).\n\nPagination par `limit` (defaut 20, max 100) et `offset` (defaut 0). Le total est inclus dans la reponse pour permettre au client d'afficher un compteur ou de preparer une pagination infinie.\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis. L'historique est strictement scope a l'utilisateur du JWT.",
        "operationId": "list_my_meal_plans_api_v1_meal_plans_me_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste pagine de plans repas (tri decroissant sur generated_at).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPlansResponse"
                },
                "example": {
                  "items": [
                    {
                      "id": 42,
                      "objective": "muscle_gain",
                      "constraints": {
                        "diet_type": "omnivore",
                        "allergies": [
                          "arachides"
                        ],
                        "duration_days": 7,
                        "budget_per_day": 15.0
                      },
                      "plan": {
                        "fallback": false,
                        "days": [
                          {
                            "day": 1,
                            "meals": [
                              {
                                "name": "Bowl quinoa, poulet et legumes verts",
                                "macros": {
                                  "calories": 620,
                                  "protein_g": 45.0,
                                  "carbs_g": 65.0,
                                  "fat_g": 18.0
                                },
                                "ingredients": [
                                  "100 g quinoa",
                                  "150 g blanc de poulet"
                                ],
                                "est_budget_eur": 5.5,
                                "prep_time_min": 25
                              }
                            ]
                          }
                        ]
                      },
                      "generated_at": "2026-04-29T10:15:00"
                    }
                  ],
                  "total": 12,
                  "limit": 20,
                  "offset": 0
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Validation Pydantic en echec (type ou contrainte de champ violee)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          }
        }
      }
    },
    "/api/v1/nutrition-goals/me": {
      "get": {
        "tags": [
          "Profil"
        ],
        "summary": "Recupere le profil nutritionnel de l'utilisateur",
        "description": "Renvoie le profil nutritionnel de l'utilisateur authentifie.\n\nLe profil regroupe l'objectif sante (`weight_loss`, `muscle_gain`, `balance`, `sport_performance`), les cibles caloriques et macros, les allergies declarees et le regime (`omnivore`, `vegetarien`, `vegan`, `sans_gluten`).\n\nRenvoie 404 si aucun profil n'a encore ete configure (un PUT initial est alors necessaire).\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis. Le profil est strictement scope a l'utilisateur du JWT.",
        "operationId": "get_my_profile_api_v1_nutrition_goals_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profil nutritionnel courant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionGoalResponse"
                },
                "example": {
                  "user_id": 42,
                  "health_goal": "muscle_gain",
                  "calories_target": 2400,
                  "protein_g": "140.0",
                  "carbs_g": "260.0",
                  "fat_g": "80.0",
                  "allergies": [
                    "arachides",
                    "fruits a coque"
                  ],
                  "diet_type": "omnivore"
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Validation Pydantic en echec (type ou contrainte de champ violee)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          },
          "404": {
            "description": "Aucun profil nutritionnel configure (faire un PUT initial)."
          }
        }
      },
      "put": {
        "tags": [
          "Profil"
        ],
        "summary": "Cree ou met a jour le profil nutritionnel (upsert)",
        "description": "Cree ou met a jour le profil nutritionnel de l'utilisateur authentifie (upsert).\n\nTous les champs sont optionnels : un PUT partiel ecrase la ligne avec la valeur fournie. Pour conserver une valeur existante, le client doit la renvoyer explicitement.\n\nLes valeurs sont utilisees ensuite par `/analyze-meal` (detection des desequilibres) et `/generate-meal-plan` (resolution de l'objectif sante par defaut).\n\n**Authentification** : header `Authorization: Bearer <jwt>` requis.",
        "operationId": "upsert_my_profile_api_v1_nutrition_goals_me_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NutritionGoalRequest"
              },
              "examples": {
                "muscle_gain": {
                  "summary": "Prise de masse, omnivore",
                  "value": {
                    "health_goal": "muscle_gain",
                    "calories_target": 2400,
                    "protein_g": 140,
                    "carbs_g": 260,
                    "fat_g": 80,
                    "allergies": [
                      "arachides"
                    ],
                    "diet_type": "omnivore"
                  }
                },
                "weight_loss_vegetarien": {
                  "summary": "Perte de poids, vegetarien",
                  "value": {
                    "health_goal": "weight_loss",
                    "calories_target": 1700,
                    "protein_g": 90,
                    "carbs_g": 170,
                    "fat_g": 55,
                    "allergies": [],
                    "diet_type": "vegetarien"
                  }
                },
                "minimal": {
                  "summary": "Profil minimal (objectif sante seul)",
                  "value": {
                    "health_goal": "balance"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil mis a jour (ou cree).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionGoalResponse"
                },
                "example": {
                  "user_id": 42,
                  "health_goal": "muscle_gain",
                  "calories_target": 2400,
                  "protein_g": "140.0",
                  "carbs_g": "260.0",
                  "fat_g": "80.0",
                  "allergies": [
                    "arachides",
                    "fruits a coque"
                  ],
                  "diet_type": "omnivore"
                }
              }
            }
          },
          "400": {
            "description": "Requete malformee (body / parametre illisible)."
          },
          "401": {
            "description": "JWT manquant, malforme ou invalide."
          },
          "403": {
            "description": "Permission insuffisante (entitlements ou tier requis pour cette operation)."
          },
          "422": {
            "description": "Payload invalide (objectif sante inconnu, valeurs non numeriques, etc)."
          },
          "503": {
            "description": "Service degrade : dependance externe (Ollama) injoignable et fallback indisponible."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DietType": {
        "type": "string",
        "enum": [
          "omnivore",
          "vegetarien",
          "vegan",
          "sans_gluten"
        ],
        "title": "DietType"
      },
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female"
        ],
        "title": "Gender"
      },
...
      "HealthGoal": {
        "type": "string",
        "enum": [
          "weight_loss",
          "muscle_gain",
          "balance",
          "sport_performance"
        ],
        "title": "HealthGoal"
      },
...
      "MealDay": {
        "properties": {
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "meals": {
            "items": {
              "$ref": "#/components/schemas/Meal"
            },
            "type": "array",
            "title": "Meals"
          }
        },
        "type": "object",
        "required": [
          "day",
          "meals"
        ],
        "title": "MealDay"
      },
...
      "MealPlanRequest": {
        "properties": {
          "health_goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HealthGoal"
              },
              {
                "type": "null"
              }
            ]
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies",
            "default": []
          },
          "budget_eur_per_day": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Eur Per Day"
          },
          "diet_type": {
            "$ref": "#/components/schemas/DietType"
          },
          "duration_days": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Duration Days",
            "default": 7
          }
        },
        "type": "object",
        "required": [
          "diet_type"
        ],
        "title": "MealPlanRequest"
      },
      "MealPlanResponse": {
        "properties": {
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "fallback": {
            "type": "boolean",
            "title": "Fallback"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/MealDay"
            },
            "type": "array",
            "title": "Days"
          },
          "compliance_status": {
            "type": "string",
            "enum": [
              "full",
              "partial_budget",
              "static_fallback"
            ],
            "title": "Compliance Status",
            "default": "full"
          },
          "compliance_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Compliance Warnings"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "fallback",
          "days"
        ],
        "title": "MealPlanResponse"
      },
...
      "NutritionGoalResponse": {
        "properties": {
          "health_goal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HealthGoal"
              },
              {
                "type": "null"
              }
            ]
          },
          "calories_target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories Target"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allergies",
            "default": []
          },
          "diet_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diet Type"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "activity_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "NutritionGoalResponse"
      },
...
      "PaginatedPlansResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MealPlanHistoryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedPlansResponse",
        "example": {
          "items": [
            {
              "constraints": {
                "allergies": [],
                "diet_type": "vegan",
                "duration_days": 7
              },
              "generated_at": "2026-04-29T10:15:00",
              "id": 42,
              "objective": "weight_loss",
              "plan": {
                "days": [
                  {
                    "day": 1,
                    "meals": []
                  }
                ]
              }
            }
          ],
          "limit": 20,
          "offset": 0,
          "total": 17
        }
      },
...
    }
  },
...
}